+++ src/main/java/com/github/tiwindetea/dungeonoflegend/events/players/inventory/InventoryDeletionEvent.java
@@ -4,7 +4,7 @@ package com.github.tiwindetea.dungeonoflegend.events.players.inventory;
  * Created by maxime on 5/6/16.
  */
 public class InventoryDeletionEvent extends InventoryEvent {
-	public InventoryDeletionEvent(byte playerNumber, int objectId) {
+	public InventoryDeletionEvent(int playerNumber, long objectId) {
 		super(playerNumber, objectId);
 	}
 }
